pyproject.toml
[project]
name = "stt-tts-unified"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "aiofiles>=23.0.0",
    "aiosqlite>=0.19.0",
    "python-multipart>=0.0.6",
    "pydantic-settings[yaml]>=2.0.0",
    "edge-tts>=6.1.0",
    "openai-whisper",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "numba>=0.59.0",
    "numpy>=1.26.0",
    "scipy",
    "ffmpeg-python==0.2.0",
    "transformers>=4.19.0",
    "pillow>=12.2.0",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cpu" }]
torchaudio = [{ index = "pytorch-cpu" }]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true
